data-model/generation-scripts/requirements.txt
# Python dependencies for CallDrop data generation scripts
cassandra-driver>=3.28.0